Add new 2023 Personal Identity Code reform characters
In 2023 the new Finnish Personal Identity Code reform introduced
new intermediate characters to PICs involving 1900s and 2000s.

This fix allows these new charactes to be identified as valid.

This fix does NOT affect the method 'getSeparatorChar' which tries to
calculate the correct intermediate character based on the date. It still
assumes only the original characters are in use for 1900s and 2000s.